% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{demodata_packages}
\alias{demodata_packages}
\title{List all packages in the demodata}
\usage{
demodata_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include demodata in the list?}
}
\description{
List all packages in the demodata
}
\examples{
demodata_packages()
}
