% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nrows_df.R
\name{get_nrows_df}
\alias{get_nrows_df}
\title{Get number of rows of a data frame}
\usage{
get_nrows_df(df)
}
\arguments{
\item{df}{data frame to analyse}
}
\value{
a message
}
\description{
Get number of rows of a data frame
}
\examples{
data("starwars_sample")
get_nrows_df(df = starwars_sample)
}
