% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver_aging_neg}
\alias{liver_aging_neg}
\title{liver aging dataset neg from MetDNA}
\format{
mass_dataset object
}
\usage{
liver_aging_neg
}
\description{
A mass_dataset object
A mass_dataset object
}
\keyword{datasets}
