% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_starwars_sample.R
\docType{data}
\name{starwars_sample}
\alias{starwars_sample}
\title{starwars_sample}
\format{
A data frame with 10 rows and 14 variables:
\describe{
\item{ name }{  name of the character }
\item{ height }{  numeric }
\item{ mass }{  numeric }
\item{ hair_color }{  character }
\item{ skin_color }{  character }
\item{ eye_color }{  character }
\item{ birth_year }{  numeric }
\item{ sex }{  character }
\item{ gender }{  character }
\item{ homeworld }{  character }
\item{ species }{  character }
\item{ films }{  character }
\item{ vehicles }{  character }
\item{ starships }{  character }
}
}
\source{
Source
}
\usage{
starwars_sample
}
\description{
A random sample of starwars dataset
}
\keyword{datasets}
